refactor: move styled folder to input folder
