+++ src/components/Input/inputBase/__test__/input.a11y.spec.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOMServer from 'react-dom/server';
 import { axe } from 'jest-axe';
-import InputBase from './../';
+import InputBase from '../';
 
 describe('<InputBase/>', () => {
     it('should be accessible when label is passed', async () => {
