+++ src/components/Input/styled/error.js
@@ -1,7 +1,7 @@
 import styled from 'styled-components';
-import { MARGIN_X_SMALL } from '../../../../styles/margins';
-import { COLOR_ERROR } from '../../../../styles/colors';
-import { FONT_SIZE_TEXT_MEDIUM } from '../../../../styles/fontSizes';
+import { MARGIN_X_SMALL } from '../../../styles/margins';
+import { COLOR_ERROR } from '../../../styles/colors';
+import { FONT_SIZE_TEXT_MEDIUM } from '../../../styles/fontSizes';
 
 const ErrorContainer = styled.div`
     font-size: ${FONT_SIZE_TEXT_MEDIUM};
