+++ src/components/Input/styled/helpText.js
@@ -1,7 +1,7 @@
 import styled from 'styled-components';
-import { MARGIN_X_SMALL } from '../../../../styles/margins';
-import { COLOR_DARK_1 } from '../../../../styles/colors';
-import { FONT_SIZE_TEXT_MEDIUM } from '../../../../styles/fontSizes';
+import { MARGIN_X_SMALL } from '../../../styles/margins';
+import { COLOR_DARK_1 } from '../../../styles/colors';
+import { FONT_SIZE_TEXT_MEDIUM } from '../../../styles/fontSizes';
 
 const HelpText = styled.div`
     font-size: ${FONT_SIZE_TEXT_MEDIUM};
