+++ src/components/Input/styled/label.js
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
-import { COLOR_GRAY_4 } from '../../../../styles/colors';
-import { FONT_SIZE_TEXT_MEDIUM } from '../../../../styles/fontSizes';
+import { COLOR_GRAY_4 } from '../../../styles/colors';
+import { FONT_SIZE_TEXT_MEDIUM } from '../../../styles/fontSizes';
 
 const Label = styled.label`
     color: ${COLOR_GRAY_4};
